requirements.txt
python-telegram-bot
wolframalpha
